+++ app/components/DBLEntryRow.js
@@ -2,6 +2,7 @@ import React, { PureComponent } from 'react';
 import { connect } from 'react-redux';
 import { compose } from 'recompose';
 import { withStyles } from '@material-ui/core/styles';
+import { Set } from 'immutable';
 import classNames from 'classnames';
 import { createSelector } from 'reselect';
 import LinearProgress from 'material-ui/LinearProgress';
@@ -16,6 +17,8 @@ import AddCircle from '@material-ui/icons/AddCircle';
 import Button from '@material-ui/core/Button';
 import FileDownload from 'material-ui/svg-icons/file/file-download';
 import Folder from '@material-ui/icons/Folder';
+import StarBorderIcon from '@material-ui/icons/StarBorder';
+import StarIcon from '@material-ui/icons/Star';
 import Copyright from '@material-ui/icons/Copyright';
 import Save from '@material-ui/icons/Save';
 import CreateNewFolder from '@material-ui/icons/CreateNewFolder';
@@ -23,6 +26,7 @@ import Description from '@material-ui/icons/Description';
 import CloudUpload from '@material-ui/icons/CloudUpload';
 import styles from './DBLEntryRow.css';
 import ControlledHighlighter from './ControlledHighlighter';
+import { toggleEntryStar } from '../actions/bundleFilter.actions';
 import { toggleSelectEntry, requestSaveBundleTo, forkIntoNewBundle,
   downloadResources, uploadBundle, updateBundle, createDraftRevision } from '../actions/bundle.actions';
 import { openEditMetadata } from '../actions/bundleEditMetadata.actions';
@@ -55,6 +59,8 @@ type Props = {
   isUploading?: ?boolean,
   isDownloading?: ?boolean,
   isSelected: ?boolean,
+  shouldShowStarred: ?boolean,
+  isStarred: ?boolean,
   shouldShowRow: boolean,
   classes: {},
   isRequestingRevision: boolean,
@@ -70,7 +76,8 @@ type Props = {
   openEditMetadata: () => {},
   uploadBundle: () => {},
   updateBundle: () => {},
-  createDraftRevision: () => {}
+  createDraftRevision: () => {},
+  toggleEntryStar: () => {}
 };
 
 const mapDispatchToProps = {
@@ -82,12 +89,14 @@ const mapDispatchToProps = {
   openEditMetadata,
   uploadBundle,
   updateBundle,
-  createDraftRevision
+  createDraftRevision,
+  toggleEntryStar
 };
 
 const getTask = (state, props) => props.task;
 const getStatus = (state, props) => props.status;
 const getIsSearchActive = (state) => state.bundlesFilter.isSearchActive;
+const getStarredEntries = (state) => state.bundlesFilter.starredEntries;
 const emptyBundleMatches = {};
 const emptyObj = {};
 const getEmptryBundleMatches = () => emptyBundleMatches;
@@ -155,9 +164,15 @@ const makeGetLaterEntryRevisions = () => createSelector(
   }
 );
 
+const makeGetIsEntryStarred = () => createSelector(
+  [getStarredEntries, getDblId],
+  (starredEntries, dblId) => (starredEntries || Set()).has(dblId)
+);
+
 const makeMapStateToProps = () => {
   const shouldShowRow = makeShouldShowRow();
   const getMatches = makeGetBundleMatches();
+  const getIsEntryStarred = makeGetIsEntryStarred();
   const getIsRequestingRevision = makeGetIsRequestingRevision();
   const getIsDownloading = makeGetIsDownloading();
   const getFormsErrors = editMetadataService.makeGetFormsErrors();
@@ -169,10 +184,12 @@ const makeMapStateToProps = () => {
       laterEntryRevisions: getLaterEntryRevisions(state, props),
       shouldShowRow: shouldShowRow(state, props),
       bundleMatches: getMatches(state, props),
+      isStarred: getIsEntryStarred(state, props),
       bundlesSaveTo,
       newMediaTypes,
       isDownloading: getIsDownloading(state, props),
-      formsErrors: getFormsErrors(state, props)
+      formsErrors: getFormsErrors(state, props),
+      shouldShowStarred: state.bundlesFilter.showStarredEntries
     };
   };
   return mapStateToProps;
@@ -217,6 +234,12 @@ class DBLEntryRow extends PureComponent<Props> {
   onClickBundleRow = () => {
     const { bundleId: id, dblId, displayAs } = this.props;
     this.props.toggleSelectEntry({ id, dblId, displayAs });
+  }
+
+  handleClickStar = (event) => {
+    const { dblId } = this.props;
+    event.stopPropagation();
+    this.props.toggleEntryStar(dblId);
   }
 
   showStatusAsText = () => {
@@ -376,7 +399,7 @@ class DBLEntryRow extends PureComponent<Props> {
   renderLicenseIcon = (license) => {
     const { classes } = this.props;
     if (license === 'owned') {
-      return <VerifiedUser className={classNames(classes.leftIcon, classes.iconSmall)} />;  
+      return <VerifiedUser className={classNames(classes.leftIcon, classes.iconSmall)} />;
     }
     if (license === 'open-access') {
       return <VerifiedUserTwoTone className={classNames(classes.leftIcon, classes.iconSmall)} />;
@@ -388,18 +411,18 @@ class DBLEntryRow extends PureComponent<Props> {
     const {
       bundleId, dblId, revision, medium, status, license,
       displayAs, progress,
-      isSelected, shouldShowRow,
+      isSelected, shouldShowRow, isStarred, shouldShowStarred,
       classes,
       newMediaTypes
     } = this.props;
     const { anchorEl } = this.state;
-    if (!shouldShowRow) {
+    if (!shouldShowRow || (shouldShowStarred && !isStarred)) {
       return (null);
     }
     const resourceManagerMode = status === 'DRAFT' ? 'addFiles' : 'download';
     const laterEntryRevisionsCount = this.props.laterEntryRevisions.length;
     const laterRevisionsBadge = laterEntryRevisionsCount ? `${laterEntryRevisionsCount}+` : '';
-    const mediumIconMarginRight = ux.getMediumIcon(medium);
+    const mediumIcon = ux.getMediumIcon(medium, { style: { marginRight: '0px' } });
     return (
       <div
         className={classNames(styles.bundleRow, this.pickBackgroundColor(true))}
@@ -411,10 +434,22 @@ class DBLEntryRow extends PureComponent<Props> {
         style={{ borderBottom: '1px solid lightgray' }}
       >
         <Grid container justify="space-between" alignItems="center" wrap="nowrap">
-          <Grid container justify="center" lg={1} md={1} sm={1}>
+          <Grid container lg={1} md={1} sm={1}>
+            <Grid item>
+              <Tooltip title="Star entry">
+                <Button size="small" style={{ minWidth: '16px' }} onClick={this.handleClickStar}>
+                  {isStarred ?
+                    <StarIcon className={classNames(classes.iconSmall)} /> :
+                    <StarBorderIcon className={classNames(classes.iconSmall)} />
+                  }
+                </Button>
+              </Tooltip>
+            </Grid>
             <Grid item>
               <Tooltip title={medium}>
-                { mediumIconMarginRight }
+                <Button size="small" style={{ minWidth: '16px' }}>
+                  { mediumIcon }
+                </Button>
               </Tooltip>
             </Grid>
           </Grid>